+++ portfolioserver/app.py
@@ -4,9 +4,7 @@ from fastapi import FastAPI
 def create_app():
     app = FastAPI()
 
-    from . import portfolio
-    from . import schemes
-    from . import goals
+    from .routes import portfolio, schemes, goals
 
     app.include_router(portfolio.router)
     app.include_router(schemes.router)
